Reorganize Internal Config Files
