+++ tools/src/config/constants.py
@@ -1,7 +1,6 @@
 import dotenv
 
-from tools.src.config import Frekvens
-from tools.src.config import Obegransad
+from . import ikeaFrekvens, ikeaObegransad
 
 ENV_FREKVENS = (
     True if "ENV_FREKVENS" in dotenv.dotenv_values("tools/.env").keys() else False
@@ -15,7 +14,15 @@ HOST = (dotenv.dotenv_values("tools/.env").get("HOSTNAME") or "frekvens") + (
 )
 
 COLUMNS = (
-    Frekvens.COLUMNS if ENV_FREKVENS else Obegransad.COLUMNS if ENV_OBEGRANSAD else None
+    ikeaFrekvens.COLUMNS
+    if ENV_FREKVENS
+    else ikeaObegransad.COLUMNS if ENV_OBEGRANSAD else 0
 )
-MODEL = Frekvens.MODEL if ENV_FREKVENS else Obegransad.MODEL if ENV_OBEGRANSAD else None
-ROWS = Frekvens.ROWS if ENV_FREKVENS else Obegransad.ROWS if ENV_OBEGRANSAD else None
+ROWS = (
+    ikeaFrekvens.ROWS if ENV_FREKVENS else ikeaObegransad.ROWS if ENV_OBEGRANSAD else 0
+)
+MODEL = (
+    ikeaFrekvens.MODEL
+    if ENV_FREKVENS
+    else ikeaObegransad.MODEL if ENV_OBEGRANSAD else None
+)
