+++ tools/src/config/ikeaFrekvens.py
