+++ webapp/src/config/ikeaFrekvens.ts
